pyproject.toml
[project]
name = "unhinged-twitter-bot"
version = "0.1.0"
description = "An unhinged Twitter bot"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "getdaft>=0.4.7",
    "kagglehub>=0.3.10",
    "lancedb>=0.21.1",
    "openai>=1.66.3",
    "pandas>=2.2.3",
    "python-dotenv>=1.0.1",
    "redis>=5.2.1",
    "sentence-transformers>=3.4.1",
    "tweepy>=4.15.0",
]

[project.scripts]
events = "scripts.events:main"
batch_ingest_subscriber = "unhinged_twitter_bot.batch_ingest_subscriber.main:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
